Animation Editor(5)
- 모든 텍스쳐 key (이름)를 파일명으로 변경.
- 프레임 추가/제거 기능 구현
- 새 애니메이션 생성 기능 구현

+++ CButtonUI.cpp
@@ -34,7 +34,7 @@ void CButtonUI::Render_DerivedUI()
 		, (int)(vFinalPos.x)
 		, (int)(vFinalPos.y)
 		, GetName().c_str()
-		, GetName().length());
+		, (int)GetName().length());
 }
 
 void CButtonUI::LBtnDown()
